+++ src/main/java/org/opentestsystem/ap/ims/rest/v1/ItemApi.java
@@ -18,6 +18,7 @@ package org.opentestsystem.ap.ims.rest.v1;
 import io.swagger.annotations.Api;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.IOUtils;
+import org.opentestsystem.ap.common.exception.ResourceNotFoundException;
 import org.opentestsystem.ap.common.exception.SystemException;
 import org.opentestsystem.ap.common.model.ItemDiff;
 import org.opentestsystem.ap.common.model.ItemHistory;
@@ -30,7 +31,11 @@ import org.opentestsystem.ap.ims.rest.ItemChangeRequest;
 import org.opentestsystem.ap.ims.rest.ItemHistoryResponse;
 import org.opentestsystem.ap.ims.service.FileUploadService;
 import org.opentestsystem.ap.ims.service.ItemService;
+import org.opentestsystem.ap.ims.service.export.ItemExportService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.FileSystemResource;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -41,10 +46,12 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletResponse;
 import javax.ws.rs.QueryParam;
+import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.List;
@@ -70,11 +77,15 @@ public class ItemApi {
 
     private final FileUploadService fileUploadService;
 
+    private final ItemExportService itemExportService;
+
     @Autowired
     public ItemApi(final ItemService itemService,
-                   final FileUploadService fileUploadService) {
+                   final FileUploadService fileUploadService,
+                   final ItemExportService itemExportService) {
         this.itemService = itemService;
         this.fileUploadService = fileUploadService;
+        this.itemExportService = itemExportService;
     }
 
     // ------------------------------------------------------------------------
@@ -364,5 +375,21 @@ public class ItemApi {
     @GetMapping("/{itemId}/resources/{section}")
     public List<String> findAttachments(@PathVariable final String itemId, @PathVariable final String section) {
         return this.fileUploadService.findAttachments(itemId, section);
+    }
+
+    @GetMapping(value = "/{itemId}/export", produces = "application/zip")
+    @ResponseBody
+    public ResponseEntity<FileSystemResource> getFile(@PathVariable("itemId") String itemId) {
+        Optional<File> maybeExport = itemExportService.exportItem(itemId);
+
+        if (!maybeExport.isPresent()) {
+            throw new ResourceNotFoundException("Could not find item " + itemId);
+        }
+
+        FileSystemResource resource = new FileSystemResource(maybeExport.get());
+        return ResponseEntity
+                .ok()
+                .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + resource.getFilename() + "\"")
+                .body(resource);
     }
 }
