--- src/main/java/org/opentestsystem/ap/ims/rest/v1/ItemExportApi.java
@@ -1,45 +1,0 @@
-package org.opentestsystem.ap.ims.rest.v1;
-
-import io.swagger.annotations.Api;
-import lombok.extern.slf4j.Slf4j;
-import org.opentestsystem.ap.common.exception.ResourceNotFoundException;
-import org.opentestsystem.ap.ims.service.export.ItemExportService;
-import org.springframework.core.io.FileSystemResource;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.io.File;
-import java.util.Optional;
-
-@Slf4j
-@RestController
-@RequestMapping("/api/v1/items")
-@Api
-public class ItemExportApi {
-    private final ItemExportService itemExportService;
-
-    public ItemExportApi(final ItemExportService itemExportService) {
-        this.itemExportService = itemExportService;
-    }
-
-    @GetMapping(value = "/{itemId}/export", produces = "application/zip")
-    @ResponseBody
-    public ResponseEntity<FileSystemResource> getFile(@PathVariable("itemId") String itemId) {
-        Optional<File> maybeExport = itemExportService.exportItem(itemId);
-
-        if (!maybeExport.isPresent()) {
-            throw new ResourceNotFoundException("Could not find item " + itemId);
-        }
-
-        FileSystemResource resource = new FileSystemResource(maybeExport.get());
-        return ResponseEntity
-                .ok()
-                .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + resource.getFilename() + "\"")
-                .body(resource);
-    }
-}
