+++ src/test/java/org/opentestsystem/ap/ims/rest/v1/ItemApiIT.java
@@ -32,6 +32,7 @@ import org.opentestsystem.ap.ims.rest.ItemChangeRequest;
 import org.opentestsystem.ap.ims.service.FileUploadService;
 import org.opentestsystem.ap.ims.service.ItemService;
 import org.opentestsystem.ap.ims.service.SavedSearchService;
+import org.opentestsystem.ap.ims.service.export.ItemExportService;
 import org.opentestsystem.ap.ims.util.IMSTestUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -122,6 +123,9 @@ public class ItemApiIT {
 
     @MockBean
     private SavedSearchService savedSearchService;
+
+    @MockBean
+    private ItemExportService itemExportService;
 
     private HttpMessageConverter mappingJackson2HttpMessageConverter;
